rename `aggregation` to `aggregate`
