+++ src/stream/src/executor/aggregate/agg_group.rs
