+++ src/stream/src/executor/aggregate/agg_state.rs
@@ -26,7 +26,7 @@ use risingwave_storage::StateStore;
 use super::minput::MaterializedInputState;
 use crate::common::StateTableColumnMapping;
 use crate::common::table::state_table::StateTable;
-use crate::executor::aggregation::agg_group::{AggStateCacheStats, GroupKey};
+use crate::executor::aggregate::agg_group::{AggStateCacheStats, GroupKey};
 use crate::executor::{PkIndices, StreamExecutorResult};
 
 /// Represents the persistent storage of aggregation state.
