+++ src/stream/src/executor/aggregate/hash_agg.rs
