+++ src/stream/src/executor/aggregate/simple_agg.rs
