+++ src/stream/src/executor/aggregate/stateless_simple_agg.rs
